requirements.txt
pytorch
transformers
nltk  # for dataset generation
telebot
question_generation